requirements.txt
Flask
pymongo
python-dotenv
PyJWT
Flask-Cors